requirements.txt
streamlit
earthengine-api
geemap
matplotlib
numpy
pandas
requests
Pillow
